data_science/ai_jobs/ProcessBattlesJob/main.py
from google.cloud import storage
import aiohttp
import asyncio
import nest_asyncio
from google.oauth2 import service_account
from googleapiclient.discovery import build
import logging
logging.getLogger().setLevel(logging.INFO)

nest_asyncio.apply()